int /> NSInteger
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3419 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/LaternaMagica/AppController.m
@@ -776,7 +776,7 @@
   NSString *destFileName;
   NSString *destFileExtension;
   NSString *destFolder;
-  int i;
+  NSInteger i;
   
   givenHeight = [fieldHeight intValue];
   givenWidth = [fieldWidth intValue];
@@ -792,7 +792,7 @@
   [exporterPanel displayIfNeeded];
   for (i = 0; i < [fileListView numberOfRows]; i++)
     {
-      int newW, newH;
+      NSInteger newW, newH;
       double aspectRatio;
       NSBitmapImageRep *scaledImageRep;
       NSAutoreleasePool *pool;
